+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixes
+  * Expect that the "resultCode" may be missing in the response
+
 ## [0.2.0] - 2023-04-02
 
 ### Added